Cleanup.
Improve variable naming and some design patterns.

+++ app/src/main/java/org/lineageos/recorder/sounds/SoundFragment.java
@@ -25,7 +25,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.ImageButton;
 import android.widget.TextView;
 
 import org.lineageos.recorder.R;
@@ -34,13 +33,16 @@ import org.lineageos.recorder.ui.SoundVisualizer;
 import org.lineageos.recorder.utils.LastRecordHelper;
 import org.lineageos.recorder.utils.Utils;
 
-public class SoundFragment extends Fragment {
+public class SoundFragment extends Fragment implements View.OnClickListener {
     private static final String TYPE = "audio/wav";
 
     private Activity mActivity;
     private TextView mCardTitle;
     private SoundVisualizer mVisualizer;
     private Button mStopButton;
+    private Button mPlayButton;
+    private Button mShareButton;
+    private Button mDeleteButton;
     private CardView mLastCard;
     private TextView mLastMessage;
 
@@ -48,35 +50,44 @@ public class SoundFragment extends Fragment {
     }
 
     @Override
-    public View onCreateView(LayoutInflater mInflater, ViewGroup mContainer,
-                             Bundle mSavedInstance) {
-        View mView = mInflater.inflate(R.layout.fragment_sound, mContainer, false);
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_sound, container, false);
 
         mActivity = getActivity();
 
-        mCardTitle = (TextView) mView.findViewById(R.id.sound_recording_title);
-        mVisualizer = (SoundVisualizer) mView.findViewById(R.id.sound_visualizer);
-        mStopButton = (Button) mView.findViewById(R.id.sound_recording_button);
-        mLastCard = (CardView) mView.findViewById(R.id.sound_card_last);
-        mLastMessage = (TextView) mView.findViewById(R.id.sound_last_message);
-        ImageButton mPlayButton = (ImageButton) mView.findViewById(R.id.sound_last_play);
-        ImageButton mShareButton = (ImageButton) mView.findViewById(R.id.sound_last_share);
-        ImageButton mDeleteButton = (ImageButton) mView.findViewById(R.id.sound_last_delete);
+        mCardTitle = (TextView) view.findViewById(R.id.sound_recording_title);
+        mVisualizer = (SoundVisualizer) view.findViewById(R.id.sound_visualizer);
+        mStopButton = (Button) view.findViewById(R.id.sound_recording_button);
+        mLastCard = (CardView) view.findViewById(R.id.sound_card_last);
+        mLastMessage = (TextView) view.findViewById(R.id.sound_last_message);
+        mPlayButton = (Button) view.findViewById(R.id.sound_last_play);
+        mShareButton = (Button) view.findViewById(R.id.sound_last_share);
+        mDeleteButton = (Button) view.findViewById(R.id.sound_last_delete);
 
-        mStopButton.setOnClickListener(mButtonView ->
-                ((RecorderActivity) mActivity).toggleSoundRecorder());
-        mPlayButton.setOnClickListener(mButtonView -> startActivityForResult(
-                LastRecordHelper.getOpenIntent(getContext(), getFile(), TYPE), 0));
-        mShareButton.setOnClickListener(mButtonView ->
-                startActivity(LastRecordHelper.getShareIntent(getContext(), getFile(), TYPE)));
-        mDeleteButton.setOnClickListener(mButtonView -> {
-            AlertDialog mDialog = LastRecordHelper.deleteFile(getContext(), getFile(), true);
-            mDialog.setOnDismissListener(mListener -> refresh(mActivity));
-            mDialog.show();
-        });
+        mStopButton.setOnClickListener(this);
+        mPlayButton.setOnClickListener(this);
+        mShareButton.setOnClickListener(this);
+        mDeleteButton.setOnClickListener(this);
 
-        refresh(getContext());
-        return mView;
+        refresh();
+        return view;
+    }
+
+    @Override
+    public void onClick(View v) {
+        if (v == mStopButton) {
+            ((RecorderActivity) getActivity()).toggleSoundRecorder();
+        } else if (v == mPlayButton) {
+            startActivityForResult(
+                    LastRecordHelper.getOpenIntent(getContext(), getFile(), TYPE), 0);
+        } else if (v == mShareButton) {
+            startActivity(LastRecordHelper.getShareIntent(getContext(), getFile(), TYPE));
+        } else if (v == mDeleteButton) {
+            AlertDialog dialog = LastRecordHelper.deleteFile(getContext(), getFile(), true);
+            dialog.setOnDismissListener(d -> refresh());
+            dialog.show();
+        }
     }
 
     public SoundVisualizer getVisualizer() {
@@ -84,35 +95,36 @@ public class SoundFragment extends Fragment {
     }
 
     // Pass context to avoid unexpected NPE when refreshing from RecorderActivity
-    public void refresh(Context mContext) {
-        if (mActivity == null) {
+    public void refresh() {
+        final Context context = getActivity();
+        if (context == null) {
             return;
         }
 
-        if (Utils.isRecording(mContext)) {
-            mCardTitle.setText(mContext.getString(Utils.isSoundRecording(mContext) ?
-                    R.string.sound_recording_title_working : R.string.sound_recording_title_busy));
-        } else {
-            mCardTitle.setText(mContext.getString(R.string.sound_recording_title_ready));
-        }
+        boolean recording = Utils.isRecording(context);
+        boolean recordingSound = Utils.isSoundRecording(context);
+        mCardTitle.setText(context.getString(
+                    recordingSound ? R.string.sound_recording_title_working :
+                    recording ? R.string.sound_recording_title_busy :
+                    R.string.sound_recording_title_ready));
+
         mVisualizer.onAudioLevelUpdated(0);
-        mStopButton.setVisibility(Utils.isSoundRecording(mContext) ? View.VISIBLE : View.GONE);
+        mStopButton.setVisibility(recordingSound ? View.VISIBLE : View.GONE);
         boolean hasLastRecord = getFile() != null;
         mLastCard.setVisibility(hasLastRecord ? View.VISIBLE : View.GONE);
         if (hasLastRecord) {
             mLastMessage.setText(mActivity.getString(R.string.sound_last_message,
-                    LastRecordHelper.getLastItemDate(mActivity, true),
-                    LastRecordHelper.getLastItemDuration(mActivity, true) / 100));
+                    LastRecordHelper.getLastItemDate(context, true),
+                    LastRecordHelper.getLastItemDuration(context, true) / 100));
         }
     }
 
     private String getFile() {
-        if (mActivity == null) {
+        final Context context = getActivity();
+        if (context == null) {
             return null;
         }
 
-        return LastRecordHelper.getLastItemPath(mActivity, true);
+        return LastRecordHelper.getLastItemPath(context, true);
     }
-
-
 }
